+++ src/components/Book/index.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import StoreProvider from './Context/Store'
-import Book from './Book'
+import Book from './Book.jsx'
 
 const ContextConection = () => {
 	return (
